+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -2,5 +2,6 @@ export 'facebook.dart';
 export 'instagram.dart';
 export 'linkedin.dart';
 export 'telegram.dart';
+export 'twitter.dart';
 export 'whatsapp.dart';
 export 'youtube.dart';